# bidforge/src/main/java/com/bidforge/controller/AuthController.java
package com.bidforge.controller;

import com.bidforge.dto.request.LoginRequest;
import com.bidforge.dto.request.RegisterRequest;
import com.bidforge.dto.response.AuthResponse;
import com.bidforge.dto.response.UserResponse;
import com.bidforge.service.AuthService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;


// Public authentication endpoints, the only POST endpoints without a token

@RestController
@RequestMapping("/api/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }


    // Create an account and return 201 Created with the new user's public profile.
    @PostMapping("/register")
    public ResponseEntity<UserResponse> register(@Valid @RequestBody RegisterRequest request) {
        UserResponse created = authService.register(request);
        return ResponseEntity.status(HttpStatus.CREATED).body(created);
    }


    // takes username and password and returns 200 OK and token and some other details
    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}