# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"BidForge and Auction Engine API","description":"\nLogin via POST /api/auth/login, click Authorize and paste the token.\nSeeded demo accounts are in companion document","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/auctions/{id}":{"get":{"tags":["auction-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}},"put":{"tags":["auction-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAuctionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auctions":{"get":{"tags":["auction-controller"],"operationId":"browse","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["SCHEDULED","OPEN","CLOSED","CANCELLED"]}},{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["ENGLISH","SEALED_BID"]}},{"name":"category","in":"query","required":false,"schema":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuctionSummaryResponse"}}}}}},"post":{"tags":["auction-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuctionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}}},"/api/auctions/{id}/open":{"post":{"tags":["auction-controller"],"operationId":"open","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}}},"/api/auctions/{id}/close":{"post":{"tags":["auction-controller"],"operationId":"close","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}}},"/api/auctions/{id}/cancel":{"post":{"tags":["auction-controller"],"operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuctionResponse"}}}}}}},"/api/auctions/{auctionId}/bids":{"get":{"tags":["bid-controller"],"operationId":"bidsOfAuction","parameters":[{"name":"auctionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBidResponse"}}}}}},"post":{"tags":["bid-controller"],"operationId":"placeBid","parameters":[{"name":"auctionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceBidRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BidResponse"}}}}}}},"/api/admin/users/{id}/status":{"patch":{"tags":["admin-controller"],"operationId":"updateUserStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/bids/my":{"get":{"tags":["bid-controller"],"operationId":"myBids","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMyBidResponse"}}}}}}},"/api/auctions/won":{"get":{"tags":["auction-controller"],"operationId":"won","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuctionSummaryResponse"}}}}}}},"/api/auctions/mine":{"get":{"tags":["auction-controller"],"operationId":"mine","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuctionSummaryResponse"}}}}}}},"/api/admin/users":{"get":{"tags":["admin-controller"],"operationId":"listUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserResponse"}}}}}}},"/api/admin/audit-events":{"get":{"tags":["admin-controller"],"operationId":"listAuditEvents","parameters":[{"name":"entityType","in":"query","required":false,"schema":{"type":"string"}},{"name":"entityId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"actor","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuditEventResponse"}}}}}}},"/api/admin/auctions":{"get":{"tags":["admin-controller"],"operationId":"listAuctions","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["SCHEDULED","OPEN","CLOSED","CANCELLED"]}},{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["ENGLISH","SEALED_BID"]}},{"name":"category","in":"query","required":false,"schema":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"seller","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuctionSummaryResponse"}}}}}}}},"components":{"schemas":{"UpdateAuctionRequest":{"type":"object","properties":{"title":{"type":"string","maxLength":100,"minLength":3},"description":{"type":"string","maxLength":2000,"minLength":0},"category":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]},"startingPrice":{"type":"number","minimum":0.01},"minIncrement":{"type":"number","minimum":0.01},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"}},"required":["category","endTime","startTime","startingPrice","title"]},"AuctionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"category":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]},"auctionType":{"type":"string","enum":["ENGLISH","SEALED_BID"]},"status":{"type":"string","enum":["SCHEDULED","OPEN","CLOSED","CANCELLED"]},"startingPrice":{"type":"number"},"minIncrement":{"type":"number"},"currentHighestBid":{"type":"number"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"sellerUsername":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"result":{"$ref":"#/components/schemas/AuctionResultResponse"}}},"AuctionResultResponse":{"type":"object","properties":{"winnerUsername":{"type":"string"},"finalPrice":{"type":"number"},"closedAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":50,"minLength":3},"email":{"type":"string","format":"email","maxLength":150,"minLength":0},"password":{"type":"string","maxLength":72,"minLength":8},"firstName":{"type":"string","maxLength":50,"minLength":0},"lastName":{"type":"string","maxLength":50,"minLength":0}},"required":["email","firstName","lastName","password","username"]},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"enabled":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"roles":{"type":"array","items":{"type":"string"}}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"tokenType":{"type":"string"},"expiresInMs":{"type":"integer","format":"int64"},"username":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"CreateAuctionRequest":{"type":"object","properties":{"title":{"type":"string","maxLength":100,"minLength":3},"description":{"type":"string","maxLength":2000,"minLength":0},"category":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]},"auctionType":{"type":"string","enum":["ENGLISH","SEALED_BID"]},"startingPrice":{"type":"number","minimum":0.01},"minIncrement":{"type":"number","minimum":0.01},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"}},"required":["auctionType","category","endTime","startTime","startingPrice","title"]},"PlaceBidRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":0.01}},"required":["amount"]},"BidResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"auctionId":{"type":"integer","format":"int64"},"bidderUsername":{"type":"string"},"amount":{"type":"number"},"createdAt":{"type":"string","format":"date-time"}}},"UpdateUserStatusRequest":{"type":"object","properties":{"enabled":{"type":"boolean"}},"required":["enabled"]},"MyBidResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"auctionId":{"type":"integer","format":"int64"},"auctionTitle":{"type":"string"},"auctionType":{"type":"string","enum":["ENGLISH","SEALED_BID"]},"auctionStatus":{"type":"string","enum":["SCHEDULED","OPEN","CLOSED","CANCELLED"]},"amount":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"currentlyWinning":{"type":"boolean"}}},"PageMyBidResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/MyBidResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"AuctionSummaryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"category":{"type":"string","enum":["ELECTRONICS","VEHICLES","ART","COLLECTIBLES","FASHION","BOOKS","SPORTS","OTHER"]},"auctionType":{"type":"string","enum":["ENGLISH","SEALED_BID"]},"status":{"type":"string","enum":["SCHEDULED","OPEN","CLOSED","CANCELLED"]},"startingPrice":{"type":"number"},"currentHighestBid":{"type":"number"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"sellerUsername":{"type":"string"}}},"PageAuctionSummaryResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AuctionSummaryResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageBidResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BidResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageUserResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"AuditEventResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"actor":{"type":"string"},"action":{"type":"string","enum":["USER_REGISTERED","USER_STATUS_CHANGED","AUCTION_CREATED","AUCTION_UPDATED","AUCTION_OPENED","AUCTION_CLOSED","AUCTION_CANCELLED","BID_PLACED","WINNER_SELECTED"]},"entityType":{"type":"string"},"entityId":{"type":"integer","format":"int64"},"details":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"PageAuditEventResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AuditEventResponse"}},"number":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}